+++ test/remote_value_tests/remote_value_sensor_test.py
@@ -3,19 +3,51 @@ import pytest
 from xknx import XKNX
 from xknx.dpt import DPTBase
 from xknx.exceptions import ConversionError
-from xknx.remote_value import RemoteValueSensor
+from xknx.remote_value import RemoteValueNumeric, RemoteValueSensor
 
 
 class TestRemoteValueSensor:
     """Test class for RemoteValueSensor objects."""
+
+    def test_value_type(self):
+        """Test initializing a value_type."""
+        xknx = XKNX()
+        assert RemoteValueSensor(xknx=xknx, value_type="pulse")
+        assert RemoteValueSensor(xknx=xknx, value_type=9)
+        assert RemoteValueSensor(xknx=xknx, value_type="9.021")
+        assert RemoteValueSensor(xknx=xknx, value_type="string")
 
     def test_wrong_value_type(self):
         """Test initializing with wrong value_type."""
         xknx = XKNX()
         with pytest.raises(ConversionError):
             RemoteValueSensor(xknx=xknx, value_type="wrong_value_type")
+            RemoteValueSensor(xknx=xknx, value_type="binary")
+            RemoteValueSensor(xknx=xknx, value_type=1)
+            RemoteValueSensor(xknx=xknx, value_type=2)
 
     def test_payload_length_defined(self):
         """Test if all members of DPTMAP implement payload_length."""
         for dpt_class in DPTBase.__recursive_subclasses__():
             assert isinstance(dpt_class.payload_length, int)
+
+
+class TestRemoteValueNumeric:
+    """Test class for RemoteValueNumeric objects."""
+
+    def test_value_type(self):
+        """Test initializing a value_type."""
+        xknx = XKNX()
+        assert RemoteValueNumeric(xknx=xknx, value_type="pulse")
+        assert RemoteValueNumeric(xknx=xknx, value_type=9)
+        assert RemoteValueNumeric(xknx=xknx, value_type="9.021")
+
+    def test_wrong_value_type(self):
+        """Test initializing with wrong value_type."""
+        xknx = XKNX()
+        with pytest.raises(ConversionError):
+            RemoteValueNumeric(xknx=xknx, value_type="string")
+        with pytest.raises(ConversionError):
+            RemoteValueNumeric(xknx=xknx, value_type=16)
+        with pytest.raises(ConversionError):
+            RemoteValueNumeric(xknx=xknx, value_type="binary")