+++ changelog.md
@@ -4,7 +4,8 @@
 
 ### Internals
 
-- fix DPTBase classmethod return types
+- Add RemoteValueNumeric for values of type `float | int`.
+- Fix DPTBase classmethod return types
 
 ## 0.18.5 DPTNumeric 2021-06-08